requirements.txt
psycopg2
django
requests
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-countries
django-money
django-filter
